+++ cli/Cargo.toml
@@ -10,6 +10,10 @@ description = "Provides the deno executable"
 repository = "https://github.com/denoland/deno"
 default-run = "deno"
 
+[lib]
+name = "deno"
+path = "lib.rs"
+
 [[bin]]
 name = "deno"
 path = "main.rs"
